+++ playground/vite.config.ts
@@ -4,10 +4,7 @@ import caddyTls from "../src/index.js";
 const config = defineConfig({
   plugins: [
     caddyTls({
-      domains: ["gg.localhost", "ok.localhost"],
-      // * as string or an array of hostnames ["domain1.example", "domain2.example"]
-      // cors: "*",
-      // other properties that we want to support
+      domains: ["ligam.localhost", "ok.localhost"],
     })
   ]
 });
